Remove dependency on nose
