+++ tests/test_concurrency/test_eventloopscheduler.py
@@ -1,5 +1,4 @@
 import unittest
-from unittest import SkipTest
 
 from datetime import datetime, timedelta
 from time import sleep
