+++ tests/test_concurrency/test_mainloopscheduler/test_asyncioscheduler.py
@@ -1,8 +1,143 @@
-from nose import SkipTest
+import asyncio
+import threading
+import unittest
+from datetime import datetime, timedelta
 
-import sys
+import rx
+from rx.concurrency import AsyncIOScheduler
 
-if sys.version_info.major < 3:
-    from .py2_asyncioscheduler import *
-else:
-    from .py3_asyncioscheduler import *
+
+class TestAsyncIOScheduler(unittest.TestCase):
+
+    def test_asyncio_schedule_now(self):
+        loop = asyncio.get_event_loop()
+        scheduler = AsyncIOScheduler(loop)
+        res = scheduler.now - datetime.now()
+        assert(res < timedelta(seconds=1))
+
+    def test_asyncio_schedule_action(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            scheduler = AsyncIOScheduler(loop)
+            ran = False
+
+            def action(scheduler, state):
+                nonlocal ran
+                ran = True
+            scheduler.schedule(action)
+
+            yield from asyncio.sleep(0.1, loop=loop)
+            assert(ran is True)
+
+        loop.run_until_complete(go())
+
+    def test_asyncio_schedule_action_threadsafe(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            scheduler = AsyncIOScheduler(loop, threadsafe=True)
+            ran = False
+
+            def action(scheduler, state):
+                nonlocal ran
+                ran = True
+
+            def schedule():
+                scheduler.schedule(action)
+
+            threading.Thread(target=schedule).start()
+
+            yield from asyncio.sleep(0.1, loop=loop)
+            assert ran is True
+
+        loop.run_until_complete(go())
+
+    def test_asyncio_schedule_action_due(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            scheduler = AsyncIOScheduler(loop)
+            starttime = loop.time()
+            endtime = None
+
+            def action(scheduler, state):
+                nonlocal endtime
+                endtime = loop.time()
+
+            scheduler.schedule_relative(200, action)
+
+            yield from asyncio.sleep(0.3, loop=loop)
+            diff = endtime-starttime
+            assert diff > 0.18
+
+        loop.run_until_complete(go())
+
+    def test_asyncio_schedule_action_due_threadsafe(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            scheduler = AsyncIOScheduler(loop, threadsafe=True)
+            starttime = loop.time()
+            endtime = None
+
+            def action(scheduler, state):
+                nonlocal endtime
+                endtime = loop.time()
+
+            def schedule():
+                scheduler.schedule_relative(200, action)
+
+            threading.Thread(target=schedule).start()
+
+            yield from asyncio.sleep(0.3, loop=loop)
+            diff = endtime-starttime
+            assert diff > 0.18
+
+        loop.run_until_complete(go())
+
+    def test_asyncio_schedule_action_cancel(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            ran = False
+            scheduler = AsyncIOScheduler(loop)
+
+            def action(scheduler, state):
+                nonlocal ran
+                ran = True
+            d = scheduler.schedule_relative(10, action)
+            d.dispose()
+
+            yield from asyncio.sleep(0.1, loop=loop)
+            assert(not ran)
+
+        loop.run_until_complete(go())
+
+    def test_asyncio_schedule_action_cancel_threadsafe(self):
+        loop = asyncio.get_event_loop()
+
+        @asyncio.coroutine
+        def go():
+            ran = False
+            scheduler = AsyncIOScheduler(loop, threadsafe=True)
+
+            def action(scheduler, state):
+                nonlocal ran
+                ran = True
+
+            def schedule():
+                d = scheduler.schedule_relative(10, action)
+                d.dispose()
+
+            threading.Thread(target=schedule).start()
+
+            yield from asyncio.sleep(0.1, loop=loop)
+            assert(not ran)
+
+        loop.run_until_complete(go())
