+++ tests/test_core/test_observer.py
@@ -1,5 +1,3 @@
-from nose.tools import assert_raises
-
 from rx.core import ObserverBase, AnonymousObserver
 from rx.core.notification import OnNext, OnError, OnCompleted, from_notifier
 from rx.internal.exceptions import CompletedException
